+++ test/test_parse2.c
@@ -1,0 +1,17 @@
+#include "minishell.h"
+
+void	put_execdata(t_execdata *data)
+{
+	int	i;
+
+	i = 0;
+	while (data)
+	{
+		printf("execdata[%d]\n", i);
+		printf("\t*status: %hhu\n", *(data->status));
+		put_cmdlist(data->clst);
+		put_iolist(data->iolst);
+		data = data->next;
+		i++;
+	}
+}
